1.0 lullaby_agent/requirements.txt
streamlit>=1.33.0
langchain>=0.1.14
langchain-google-genai>=0.0.7
python-dotenv>=1.0.1
google-generativeai>=0.3.2
